Picks the largest profit
We knew right from the start this path wouldn’t lead to an efficient solution.
Now we are making not one but two recursive calls!

Moreover there is yet another problem as serious as the efficiency. What should
happen if we get a negative stock price? What if we get the string “Pssst… wanna
grab some free NFTs?”. Not only we are not validating the data, we are not using
objects to model the domain. Prices are not integers; they should at least be
associated with a currency. We are using Ruby as if it were Pascal or some other
structured-programming language.

This is evidence that —although it can help— test-driven development does not
guarantee a good design! The purpose for doing this first implementation is to
be able to profile it and compare it to the next one.

See you soon!